code for replacing specific text in a file
Try hiding the words 'javascript' and 'php' in a text file named 'test.txt'. Run code to see text replaced to 'Java' and 'python'. Haven't tested this yet but I think you can run this anywhere as long as your replace 'test.txt' in code with the path to test.txt.